Refactor GitHub API Explorer
