+++ app/controllers/github_explorer/api_requests_controller.rb
@@ -3,7 +3,6 @@
 module GithubExplorer
   class ApiRequestsController < ApplicationController
     before_action :set_new_params, only: %i[new]
-    before_action :set_octokit_client, only: %i[create]
     before_action :set_github_explorer_api_request, only: %i[show destroy]
 
     def index
@@ -14,17 +13,13 @@ module GithubExplorer
 
     def new
       @github_explorer_api_request = GithubExplorer::ApiRequest.new(@new_params)
-      @resource_options = GithubExplorer::ApiRequest.resources.map do |k, v|
-        [k.titleize, GithubExplorer::ApiRequest.resources.key(v)]
-      end
     end
 
     def create
-      response = get_response_by_resource(github_explorer_api_request_params)
-      create_params = github_explorer_api_request_params.merge(user: current_user, response:)
+      create_params = api_request_params.merge(user: current_user)
       @github_explorer_api_request = GithubExplorer::ApiRequest.new(create_params)
 
-      if @github_explorer_api_request.save
+      if @github_explorer_api_request.send_request_and_save!
         redirect_to github_explorer_api_request_url(@github_explorer_api_request)
       else
         render :new, status: :unprocessable_entity
@@ -43,7 +38,7 @@ module GithubExplorer
       @github_explorer_api_request = GithubExplorer::ApiRequest.find(params[:id])
     end
 
-    def github_explorer_api_request_params
+    def api_request_params
       params.require(:github_explorer_api_request).permit(:resource, :query, :url)
     end
 
@@ -51,23 +46,6 @@ module GithubExplorer
       return unless params[:url]
 
       @new_params = { url: params[:url], resource: 'url' }
-    end
-
-    def set_octokit_client
-      @client = Octokit::Client.new(access_token: current_user.github_token, per_page: 100)
-    end
-
-    def get_response_by_resource(params)
-      case params[:resource]
-      when 'url'
-        @client.get(params[:url])
-      when 'repositories'
-        @client.repositories
-      when 'organizations'
-        @client.organizations
-      when 'user'
-        @client.user
-      end
     end
   end
 end
