+++ app/helpers/github_explorer/api_requests_helper.rb
@@ -2,92 +2,55 @@
 
 module GithubExplorer
   module ApiRequestsHelper
-    # ============================
-    # Helpers
-    # ============================
-    def array?(obj)
-      obj.is_a?(Array) && obj.length > 2
+    include JsonHelper
+
+    def resource_options
+      GithubExplorer::ApiRequest.resources.map do |k, v|
+        label = k == 'url' ? "#{k.titleize} (must specify url)" : k.titleize
+        [label, GithubExplorer::ApiRequest.resources.key(v)]
+      end
     end
 
-    def obj?(obj)
-      obj.is_a?(Array) && obj.length == 2
+    def tag_show_item(api_request)
+      tag.div class: 'mt-6' do
+        concat tag_list_item_attrs(api_request)
+        concat tag.pre render_api_response(api_request.response), class: 'bg-gray-600 p-5'
+      end
     end
 
-    def rel?(key, val)
-      key.end_with?('url') && val&.include?('api.github.com')
+    def tag_list_item(api_request)
+      tag.div class: 'mt-6' do
+        concat tag_list_item_attrs(api_request)
+        concat tag_list_item_links(api_request)
+        concat tag.hr class: 'mt-6'
+      end
     end
 
-    def depth(arr)
-      return 0 unless arr.is_a?(Array)
-
-      1 + depth(arr[0])
+    def tag_list_item_attrs(api_request)
+      tag.div class: 'mt-6' do
+        concat tag.div "ID: #{api_request.id}"
+        concat tag.div "Requested by: #{api_request.user.github_username}"
+        concat tag_resource_or_url(api_request)
+        concat tag.div "Created: #{api_request.created_at}"
+      end
     end
 
-    # ============================
-    # Rendering
-    # ============================
-    def render_api_response(response)
-      if depth(response) > 2
-        render_json_array(response, -1)
+    def tag_resource_or_url(api_request)
+      if api_request.url?
+        tag.div "Url: #{api_request.url}"
       else
-        render_json_object(response, -1)
+        tag.div "Resource: #{api_request.resource.titleize}"
       end
     end
 
-    def render_json_object(obj, depth)
-      if obj?(obj)
-        obj_content(obj, depth)
-      elsif array?(obj)
-        render_json_array(obj, depth + 1)
+    def tag_list_item_links(api_request)
+      tag.div class: 'flex justify-between mt-6' do
+        concat link_to 'Show', api_request, class: 'text-blue-600'
+        concat button_to 'Delete',
+                         github_explorer_api_request_path(api_request),
+                         method: :delete,
+                         class: 'text-red-600'
       end
-    end
-
-    def render_json_array(arr, depth)
-      content_tag(:div) do
-        arr.each do |obj|
-          concat render_json_object(obj, depth)
-        end
-      end
-    end
-
-    # ============================
-    # Content tags
-    # ============================
-    def obj_content(obj, depth)
-      key = obj[0]
-      val = obj[1]
-      content_tag(:div) do
-        concat "#{key}: ".indent(depth * 4)
-        concat val_content(val, depth)
-
-        rel?(key, val) && (concat rel_link(key, val))
-      end
-    end
-
-    def val_content(val, depth)
-      if array?(val)
-        render_json_array(val, depth + 1)
-      else
-        val
-      end
-    end
-
-    def rel_link(key, val)
-      rel = rel_parts(key, val)
-      return unless rel[:params].length <= 1
-
-      link_to "GET #{rel[:name]}",
-              new_github_explorer_api_request_path(url: rel[:url]),
-              class: 'ml-2 rounded-md bg-gray-100 text-blue-600'
-    end
-
-    def rel_parts(key, val)
-      parts = val.split('{/')
-      {
-        name: key.gsub('_url', ''),
-        url: parts[0],
-        params: parts[1..].map { |p| p.gsub('}', '') }
-      }
     end
   end
 end
