+++ app/views/github_explorer/api_requests/_api_request.html.erb
@@ -1,22 +1,1 @@
-<div id="<%= dom_id api_request %>", class="mt-6">
-  <p class="my-5">
-    <strong class="font-medium mb-1">Created:</strong>
-    <%= api_request.created_at %>
-  </p>
-
-  <p class="my-5">
-    <strong class="font-medium mb-1">Resource:</strong>
-    <%= api_request.resource %>
-  </p>
-
-  <p class="my-5">
-    <strong class="font-medium mb-1">Url:</strong>
-    <%= api_request.url %>
-  </p>
-
-  <div class="flex justify-between">
-    <%= link_to "Show Response", api_request, class: "text-blue-600" %>
-    <%= button_to 'Delete Request', github_explorer_api_request_path(api_request), method: :delete, class: "text-red-600" %>
-  </div>
-  <hr class="mt-6">
-</div>
+<%= tag_list_item(api_request) %>
