+++ app/views/github_explorer/api_requests/_form.html.erb
@@ -13,7 +13,7 @@
 
   <div class="mt-6">
     <%= form.label :resource %>
-    <%= form.select :resource, options_for_select(@resource_options, @github_explorer_api_request.resource) %>
+    <%= form.select :resource, options_for_select(resource_options, @github_explorer_api_request.resource) %>
   </div>
 
   <div class="mt-6">
